# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Documentación API del Sistema de Gestión de Habilitaciones Municipales de Rauch'
  description: 'Sistema de digitalización de expedientes de habilitaciones comerciales de la Municipalidad de Rauch'
  version: 1.0.0
paths:
  /login:
    post:
      tags:
        - auth
      summary: 'Iniciar sesión en el sistema'
      description: 'Logueo mediante email y password'
      operationId: authLogin
      requestBody:
        description: 'Credenciales del ususario'
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: usuario@mail.com
                password:
                  type: string
                  format: password
                  example: PassWord12345
                remember:
                  type: boolean
                  example: 'true'
              type: object
      responses:
        200:
          description: 'Credencial válida'
          content:
            application/json:
              schema:
                properties:
                  message: []
                type: string
              example: 'Usuario válido.'
        422:
          description: 'Error en la credencial'
          content:
            application/json:
              schema:
                properties:
                  message: []
                type: string
              example: 'Usuario no válido.'
  /loguot:
    post:
      tags:
        - auth
      summary: 'Terminar sesión en el sistema'
      description: 'Salir del sistema.'
      operationId: authLogout
      responses:
        200:
          description: 'Sesión terminada'
          content:
            application/json:
              schema:
                properties:
                  message: []
                type: string
              example: 'Sesión terminada.'
  /index:
    get:
      tags:
        - index
      summary: 'Página principal del sistema.'
      description: 'Ingreso al sistema luego del inicio de sesión.'
      operationId: index
      responses:
        200:
          description: 'Página principal'